Move `Op` to `internal` package
