+++ core/src/main/scala/pink/cozydev/lucille/internal/Op.scala
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package pink.cozydev.lucille
+package pink.cozydev.lucille.internal
 
 import cats.data.NonEmptyList
+import pink.cozydev.lucille.Query
 
 sealed trait Op extends Product with Serializable
